Moves getContentUrls to the DomainAccessManager
