+++ domain_source/domain_source.services.yml
@@ -4,6 +4,6 @@ services:
     arguments: ['@entity_type.manager']
   domain_source.path_processor:
     class: Drupal\domain_source\HttpKernel\DomainSourcePathProcessor
-    arguments: ['@domain.negotiator', '@domain_access.manager', '@module_handler', '@entity_type.manager', '@path.alias_manager', '@config.factory']
+    arguments: ['@domain.negotiator', '@module_handler', '@entity_type.manager', '@path.alias_manager', '@config.factory']
     tags:
       - { name: path_processor_outbound, priority: 90 }
